refactor(console): split server.py monolith into console package
Extract the 1,708-line data/server.py Flask monolith into a data/console
package with a create_app() factory, eight per-area blueprints (home,
poll_import, polls, westminster, holyrood, exports, by_elections, db_admin),
and shared services: a single subprocess runner + command_result renderer,
the import preview cache, trend-cache loading, and the model-output
aggregation parameterised by ElectionType (ready for Holyrood reuse).

Templates and static assets move under console/ (git-tracked renames) and
url_for calls are re-pointed to blueprint endpoints. server.py is now a thin
entrypoint so `python server.py` still works.

No behaviour change: all 22 routes preserved, 246 tests pass.
