+++ data/console/blueprints/__init__.py
@@ -1,0 +1,1 @@
+"""Flask blueprints for the console web app, one per area of functionality."""
