+++ data/console/services/__init__.py
@@ -1,0 +1,1 @@
+"""Shared service helpers for the console web app."""
