+++ data/console/templates/base.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <header class="topbar">
-      <a href="{{ url_for('home') }}" class="brand">Election Data Console</a>
+      <a href="{{ url_for('home.home') }}" class="brand">Election Data Console</a>
     </header>
 
     <main class="container">
