+++ data/console/templates/command_result.html
