+++ data/console/templates/home.html
@@ -5,27 +5,27 @@
     <p>Run poll imports, refresh polling data, and manage model runs and outputs.</p>
 
     <div class="action-grid action-grid-three">
-      <a class="button primary button-block" href="{{ url_for('import_poll_form') }}">Import Poll</a>
-      <form method="post" action="{{ url_for('update_polls') }}" class="inline-form button-form">
+      <a class="button primary button-block" href="{{ url_for('poll_import.import_poll_form') }}">Import Poll</a>
+      <form method="post" action="{{ url_for('westminster.update_polls') }}" class="inline-form button-form">
         <button type="submit" class="button button-block">Sync from Wikipedia</button>
       </form>
-      <a class="button button-block" href="{{ url_for('poll_list') }}">View Polls</a>
+      <a class="button button-block" href="{{ url_for('polls.poll_list') }}">View Polls</a>
     </div>
 
     <div class="action-grid action-grid-two">
-      <a class="button button-block" href="{{ url_for('model_run_form') }}">Run Model</a>
-      <a class="button button-block" href="{{ url_for('model_outputs') }}">View Model Outputs</a>
+      <a class="button button-block" href="{{ url_for('westminster.model_run_form') }}">Run Model</a>
+      <a class="button button-block" href="{{ url_for('westminster.model_outputs') }}">View Model Outputs</a>
     </div>
 
     <div class="action-grid action-grid-two">
-      <a class="button button-block" href="{{ url_for('by_election_form') }}">Import By-Election</a>
-      <form method="post" action="{{ url_for('export_current_simulation') }}" class="inline-form button-form">
+      <a class="button button-block" href="{{ url_for('by_elections.by_election_form') }}">Import By-Election</a>
+      <form method="post" action="{{ url_for('exports.export_current_simulation') }}" class="inline-form button-form">
         <button type="submit" class="button button-block">Export Simulation JSON</button>
       </form>
     </div>
 
     <div class="action-grid action-grid-two">
-      <form method="post" action="{{ url_for('holyrood_import_polls') }}" class="inline-form button-form">
+      <form method="post" action="{{ url_for('holyrood.holyrood_import_polls') }}" class="inline-form button-form">
         <button type="submit" class="button button-block">Import Scottish Polls</button>
         <small>
           Constituency: {{ holyrood_latest_constituency or '—' }}
@@ -36,11 +36,11 @@
     </div>
 
     <div class="action-grid action-grid-two">
-      <form method="post" action="{{ url_for('backup_database') }}" class="inline-form button-form">
+      <form method="post" action="{{ url_for('db_admin.backup_database') }}" class="inline-form button-form">
         <button type="submit" class="button button-block">Backup Database</button>
         <small>Snapshot the local database to Google Drive.</small>
       </form>
-      <form method="post" action="{{ url_for('restore_database') }}" class="inline-form button-form"
+      <form method="post" action="{{ url_for('db_admin.restore_database') }}" class="inline-form button-form"
             onsubmit="return confirm('Restore the database from the Google Drive snapshot? This overwrites the current local database (a .prerestore copy is kept).');">
         <button type="submit" class="button button-block button-danger">Sync from Database</button>
         <small>Restore the local database from the Google Drive snapshot.</small>
