+++ data/console/templates/model_run.html
@@ -10,7 +10,7 @@
 
   <section class="card">
     <h2>Run UNS Model</h2>
-    <form method="post" action="{{ url_for('model_run_execute') }}" class="form">
+    <form method="post" action="{{ url_for('westminster.model_run_execute') }}" class="form">
       <label for="map_name">{{ map_name_help }} <span class="arg-flag">(--map-name)</span></label>
       <select name="map_name" id="map_name" required>
         {% for map_name in choices.map_names %}
