+++ data/console/templates/poll_detail.html
@@ -13,7 +13,7 @@
         -
       {% endif %}
     </p>
-    <p><a class="button" href="{{ url_for('poll_detail_csv', poll_id=poll.id) }}">Download CSV</a></p>
+    <p><a class="button" href="{{ url_for('polls.poll_detail_csv', poll_id=poll.id) }}">Download CSV</a></p>
   </section>
 
   <section class="card">
